+++ trilby-cli/src/Trilby/HNix.hs
@@ -5,15 +5,17 @@
 
 module Trilby.HNix where
 
+import Data.ByteString.Lazy qualified as LazyByteString
 import Data.Fix
 import Data.List.Extra qualified as List
 import Data.List.NonEmpty qualified as NonEmpty
 import Data.Text qualified as Text
+import Data.Text.Encoding qualified as Text
 import Lens.Family.TH (makeTraversals)
 import Nix
 import Nix.Atoms (NAtom (NNull))
 import Trilby.Host
-import Turtle qualified
+import Trilby.Process
 import Prelude
 
 instance IsString (NAttrPath NExpr) where
@@ -74,9 +76,9 @@ data FileOrFlake
 
 nixBuild :: (HasCallStack) => FileOrFlake -> App [Path Abs Dir]
 nixBuild f =
-    mapM (parseAbsDir . Text.unpack) . Text.lines
+    mapM (parseAbsDir . fromText) . Text.lines . Text.decodeUtf8 . LazyByteString.toStrict
         =<< withTrace
-            cmd
+            (readProcessStdout_ . proc)
             ( sconcat
                 [ ["nix", "build"]
                 , ["--no-link", "--print-out-paths"]
@@ -89,8 +91,7 @@ nixBuild f =
 copyClosure :: (HasCallStack) => Host -> Path Abs Dir -> App ()
 copyClosure Localhost _ = pure ()
 copyClosure host@Host{} path = do
-    (code, _) <- ssh host cmd' ["command", "-v", "nix-store"]
-    case code of
+    ssh host cmdCode ["command", "-v", "nix-store"] >>= \case
         ExitSuccess ->
             cmd_ . sconcat $
                 [ ["nix-copy-closure"]
@@ -100,8 +101,8 @@ copyClosure host@Host{} path = do
                 , [fromPath path]
                 ]
         _ -> do
-            ssh host rawCmd_ ["sudo mkdir -p /nix && sudo chown -R $(whoami) /nix"]
-            flip shell_ Turtle.stdin . Text.intercalate " " . sconcat $
+            ssh host shell_ ["sudo mkdir -p /nix && sudo chown -R $(whoami) /nix"]
+            shell_ . sconcat $
                 [
                     [ "nix-store"
                     , "--query"
@@ -119,7 +120,7 @@ copyClosure host@Host{} path = do
                     ]
                 , ["|"]
                 , sconcat
-                    [ "ssh" : ["-t", ishow host]
+                    [ "ssh" :| ["-t", ishow host]
                     ,
                         [ "tar"
                         , "--extract"
@@ -132,5 +133,5 @@ copyClosure host@Host{} path = do
 
 trilbyFlake :: (HasCallStack) => [Text] -> App FlakeRef
 trilbyFlake output = do
-    hasTrilby <- (ExitSuccess ==) . fst <$> cached quietCmd' ["nix", "flake", "metadata", "trilby"]
+    hasTrilby <- (ExitSuccess ==) <$> cached cmdCode ["nix", "flake", "metadata", "trilby"]
     pure FlakeRef{url = if hasTrilby then "trilby" else "github:ners/trilby", output}